requirements.txt
# Core scientific stack
numpy>=1.26
pandas>=2.1
scipy>=1.11
scikit-learn>=1.3
matplotlib>=3.8
plotly>=5.18

# Live data
yfinance>=0.2.40        # market data (free)
fredapi>=0.5.1          # FRED macro data
requests>=2.31          # generic HTTP (NewsAPI, fallback)

# Optional paid/expanded providers (install only if used)
# polygon-api-client>=1.13
# alpaca-py>=0.20

# UI
streamlit>=1.31
streamlit-autorefresh>=1.0.1
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.7

# Excel report generation
openpyxl>=3.1

# Utility
python-dotenv>=1.0
pyyaml>=6.0
joblib>=1.3

# Local database and validation helpers use stdlib sqlite3
